fix(migrating from GSON to Moshi)
- Migrating Retrofit's converter from GSON to Moshi.

+++ app/src/main/java/com/rudio/moviesdemo/ui/BindingAdapters.kt
@@ -19,7 +19,7 @@ import com.rudio.moviesdemo.utils.PicassoHelper
 import com.rudio.moviesdemo.utils.TextViewHelper
 
 @BindingAdapter("android:text")
-fun setText(textView: TextView, value: Float) {
+fun setText(textView: TextView, value: Double) {
     textView.text = value.toString()
 }
